422000326204-Website-ban-laptop/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
sqlalchemy==2.0.32
asyncpg==0.29.0
motor==3.5.1
PyJWT==2.9.0
